Tests/LibWeb/Text/input/async-scrolling/rounded-overlay-wheel-target.html

<!DOCTYPE html>
<script src="../include.js"></script>
<style>
    body {
        margin: 0;
    }

    #scroller {
        position: absolute;
        left: 0;
        top: 0;
        width: 200px;
        height: 200px;
        overflow: scroll;
    }

    #spacer {
        height: 800px;
    }

    #overlay {
        position: absolute;
        z-index: 1;
        left: 0;
        top: 0;
        width: 100px;
        height: 100px;
        border-radius: 50px;
        background-color: red;
    }

    #page-spacer {
        height: 2000px;
        margin-top: 200px;
    }
</style>
<div id="scroller"><div id="spacer"></div></div>
<div id="overlay"></div>
<div id="page-spacer"></div>
<script>
    test(() => {
        println(`rounded corner wheel target: ${internals.asyncScrollingStateWheelTargetAt(5, 5, 0, 100)}`);
        println(`rounded center wheel target: ${internals.asyncScrollingStateWheelTargetAt(50, 50, 0, 100)}`);
    });
</script>
