Tests/LibWeb/Text/expected/async-scrolling/rounded-overlay-wheel-target.txt

rounded corner wheel target: non-viewport
rounded center wheel target: viewport
